+++ CHANGELOG.md
@@ -1,3 +1,19 @@
+# [1.2.0](https://github.com/commown-shared-wallet/contracts-dapps/compare/v1.1.0...v1.2.0) (2022-05-15)
+
+
+### Features
+
+* **contract-types:** typing for contract, change typechain configuration ([0d16c4d](https://github.com/commown-shared-wallet/contracts-dapps/commit/0d16c4d9dfe93217bb3dd7cb2dd57c7aa50a86a1))
+* **contract-types:** typing for contract, change typechain configuration ([3a2ade8](https://github.com/commown-shared-wallet/contracts-dapps/commit/3a2ade8722348ede07619376d84896ee34d1132e))
+* **contract-types:** typing for contract, change typechain configuration ([7791cff](https://github.com/commown-shared-wallet/contracts-dapps/commit/7791cff41e640f15ef571b3d46f257ff6a2c9d53))
+* **create-csw:** page to create CSW ([3758e22](https://github.com/commown-shared-wallet/contracts-dapps/commit/3758e22fe41b68859f86261b77fe24939040610b))
+* **create-csw:** page to create CSW ([f17547c](https://github.com/commown-shared-wallet/contracts-dapps/commit/f17547c04d56371d73a282cca06cc3c434ca11d4))
+* **creation-csw:** creation of shared wallet ([f93f736](https://github.com/commown-shared-wallet/contracts-dapps/commit/f93f73605972f369cda2bfede3bff85d473f8bb7))
+* **creation-csw:** creation of shared wallet ([df23c61](https://github.com/commown-shared-wallet/contracts-dapps/commit/df23c6163456428abe8e12d662706931d9e54a4d))
+* **creation-csw:** creation of shared wallet ([5f390a0](https://github.com/commown-shared-wallet/contracts-dapps/commit/5f390a02bbaac112920912f0cb762aba1950c012))
+* **getContract:** add CommownSW contract in frontend ([c1258f1](https://github.com/commown-shared-wallet/contracts-dapps/commit/c1258f192f4cba25e86bdc8a6932108720a18835))
+* **getContract:** add CommownSW contract in frontend ([4656ed9](https://github.com/commown-shared-wallet/contracts-dapps/commit/4656ed9c312ba0156104290cf7a2ded9ba680ade))
+
 # [1.1.0](https://github.com/commown-shared-wallet/contracts-dapps/compare/v1.0.0...v1.1.0) (2022-05-08)